--- NatureUp/Views/HomeView.swift
@@ -1,20 +1,0 @@
-//
-//  HomeView.swift
-//  NatureUp
-//
-//  Created by Wonhyuk Choi on 2022/04/06.
-//
-
-import SwiftUI
-
-struct HomeView: View {
-    var body: some View {
-        Text(/*@START_MENU_TOKEN@*/"Hello, World!"/*@END_MENU_TOKEN@*/)
-    }
-}
-
-struct HomeView_Previews: PreviewProvider {
-    static var previews: some View {
-        HomeView()
-    }
-}
